Initial Commit
So far I've only added three Forms;
 - Login/ Register Page
 - Dev Panel
 - Slot Machine Page

Started to transfer code from the Dice Roll Game to modify to work with the Slot Machine.

Built a basic Database SQL on my Windows desktop which has a single Login on it but I'm currently just using Login details I've inputed into the code.

------

Need to figure out how to call information from the SQL Database and add information into that Database which can be associated with a users Login Details.

Need to figure out a way to hide the Dev Menu on the login screen and make a way so it can be visible to Dev's Via a keybind or something?